AMBARI-13683: Express Upgrade: HBASE upgrade fails during EU for HDP-2.1-> HDP-2.3 (jluniya)

+++ ambari-server/src/main/resources/common-services/HBASE/0.96.0.2.0/package/scripts/hbase_service.py
@@ -43,6 +43,7 @@ def hbase_service(
 
       Execute ( daemon_cmd,
         user = params.hbase_user,
+        only_if = no_op_test,
         # BUGFIX: hbase regionserver sometimes hangs when nn is in safemode
         timeout = 30,
         on_timeout = format("! ( {no_op_test} ) || {sudo} -H -E kill -9 `cat {pid_file}`"),
